+++ geopackage-sdk/src/main/java/mil/nga/geopackage/db/GeoPackageConnection.java
@@ -77,12 +77,78 @@ public class GeoPackageConnection extends GeoPackageCoreConnection {
         if (where != null) {
             countQuery.append(" where ").append(where);
         }
+        String sql = countQuery.toString();
 
-        Cursor countCursor = db.rawQuery(countQuery.toString(), args);
-        countCursor.moveToFirst();
-        int count = countCursor.getInt(0);
-        countCursor.close();
+        int count = singleResultQuery(sql, args);
+
         return count;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Integer min(String table, String column, String where, String[] args) {
+
+        Integer min = null;
+        if (count(table, where, args) > 0) {
+            StringBuilder minQuery = new StringBuilder();
+            minQuery.append("select min(").append(column).append(") from ")
+                    .append(table);
+            if (where != null) {
+                minQuery.append(" where ").append(where);
+            }
+            String sql = minQuery.toString();
+
+            min = singleResultQuery(sql, args);
+        }
+
+        return min;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Integer max(String table, String column, String where, String[] args) {
+
+        Integer max = null;
+        if (count(table, where, args) > 0) {
+            StringBuilder maxQuery = new StringBuilder();
+            maxQuery.append("select max(").append(column).append(") from ")
+                    .append(table);
+            if (where != null) {
+                maxQuery.append(" where ").append(where);
+            }
+            String sql = maxQuery.toString();
+
+            max = singleResultQuery(sql, args);
+        }
+
+        return max;
+    }
+
+    /**
+     * Query the SQL for a single integer result
+     *
+     * @param sql
+     * @param args
+     * @return int result
+     */
+    private int singleResultQuery(String sql, String[] args) {
+
+        Cursor countCursor = db.rawQuery(sql, args);
+
+        int result = 0;
+        try {
+            if (countCursor.moveToFirst()) {
+                result = countCursor.getInt(0);
+            }
+        } finally {
+            countCursor.close();
+        }
+
+        return result;
     }
 
     /**